services/ai_assistant/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
google-generativeai>=0.8.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
